+++ lib/schemas/message_shema_groups.dart
@@ -1,11 +1,13 @@
 import 'package:uuid/uuid.dart';
 
-Map<String, dynamic> messageSchemaGroup(
-    {required String content,
-    required bool isSender,
-    required List<Map<String, dynamic>> reaction,
-    required String autoMessageId,
-    required List<Map<String, dynamic>> users}) {
+Map<String, dynamic> messageSchemaGroup({
+  required String content,
+  required bool isSender,
+  required List<Map<String, dynamic>> reaction,
+  required String autoMessageId,
+  required List<Map<String, dynamic>> users,
+  List<Map<String, String>>? attachments, // New parameter for attachments
+}) {
   return {
     "content": {
       "autoMessageId": autoMessageId,
@@ -28,6 +30,7 @@ Map<String, dynamic> messageSchemaGroup(
               "distributed": true, // Default value for distributed
             })
         .toList(),
+    "attachments": attachments ?? [], // Add attachments to the message
   };
 }